Рефакторинг кода

+++ src/views/IndividualWorks/stores/individualWork.js
@@ -4,11 +4,11 @@ import { useVuelidate } from "@vuelidate/core";
 import individualWorkFormRules from "@/validators/individualWorkFormRules.js";
 import ky from "ky";
 import { API_URL } from "@/constants";
-import { useConfirm } from "primevue/useconfirm";
 import DialogForm from "@/utils/dialog";
-import { useToast } from "primevue/usetoast";
+import { useConfirmStore } from "@/stores/confirms";
 
 export const useIndividualWorkStore = defineStore("individualWork", () => {
+  const { confirmAdd, confirmEdit, confirmDelete } = useConfirmStore();
   const dialog = ref(
     new DialogForm({
       add: "Добавление записи",
@@ -16,9 +16,6 @@ export const useIndividualWorkStore = defineStore("individualWork", () => {
       view: "Просмотр записи",
     })
   );
-
-  const toast = useToast();
-  const confirm = useConfirm();
 
   const columns = ref([
     { field: "workDate", header: "Дата" },
@@ -54,30 +51,20 @@ export const useIndividualWorkStore = defineStore("individualWork", () => {
   const isSubmit = ref(false);
   async function addIndividualWork() {
     isSubmit.value = true;
-    if (!v$.value.$invalid) {
-      const createdIndividualWork = await ky
-        .post(`${API_URL}/individualWorks`, {
-          json: { studentId: studentId.value, ...individualWork.value },
-        })
-        .json();
-      individualWorks.value.push(createdIndividualWork);
-      setEmptyIndividualWork();
-      isSubmit.value = false;
-      dialog.value.closeDialog();
-      toast.add({
-        severity: "success",
-        summary: "Успех",
-        detail: "Запись успешно добавлена",
-        life: 2000,
-      });
-    } else {
-      toast.add({
-        severity: "warn",
-        summary: "Предупреждение",
-        detail: "Заполните все обязательные поля",
-        life: 2000,
-      });
-    }
+    confirmAdd({
+      invalid: v$.value.$invalid,
+      funcIf: async () => {
+        const createdIndividualWork = await ky
+          .post(`${API_URL}/individualWorks`, {
+            json: { studentId: studentId.value, ...individualWork.value },
+          })
+          .json();
+        individualWorks.value.push(createdIndividualWork);
+        setEmptyIndividualWork();
+        isSubmit.value = false;
+        dialog.value.closeDialog();
+      },
+    });
   }
 
   async function editIndividualWork() {
@@ -85,51 +72,30 @@ export const useIndividualWorkStore = defineStore("individualWork", () => {
       json: individualWork.value,
     });
 
-    toast.add({
-      severity: "info",
-      summary: "Инфо",
-      detail: "Запись успешно изменена",
-      life: 2000,
-    });
-
     await fetchIndividualWorks(studentId.value);
   }
 
-  const confirmEditIndividualWork = () =>
-    confirm.require({
-      message: "Вы точно хотите изменить запись?",
-      header: "Изменение записи",
-      icon: "pi pi-info-circle",
-      rejectLabel: "Отмена",
-      acceptLabel: "Изменить",
-      rejectClass: "p-button-secondary p-button-outlined",
-      acceptClass: "p-button-info",
-      accept: () => {
-        if (v$.value.$invalid) {
-          isSubmit.value = true;
-          toast.add({
-            severity: "warn",
-            summary: "Предупреждение",
-            detail: "Заполните все обязательные поля",
-            life: 2000,
-          });
-          return;
-        }
-        editIndividualWork();
+  const confirmEditIndividualWork = () => {
+    isSubmit.value = true;
+    confirmEdit({
+      invalid: v$.value.$invalid,
+      funcAccept: async () => {
+        await editIndividualWork();
         dialog.value.closeDialog();
         isSubmit.value = false;
       },
-      reject: () => {
+      funcReject: async () => {
         setEmptyIndividualWork();
         dialog.value.closeDialog();
       },
     });
+  };
 
   const selectedIndividualWorks = ref([]);
 
   function deleteSelectedIndividualWorks() {
-    selectedIndividualWorks.value.forEach((individualWork) => {
-      ky.delete(`${API_URL}/individualWorks/${individualWork.id}`);
+    selectedIndividualWorks.value.forEach(async (individualWork) => {
+      await ky.delete(`${API_URL}/individualWorks/${individualWork.id}`);
       individualWorks.value = individualWorks.value.filter(
         (work) => work.id !== individualWork.id
       );
@@ -138,24 +104,11 @@ export const useIndividualWorkStore = defineStore("individualWork", () => {
   }
 
   const confirmDeleteIndividualWorks = () =>
-    confirm.require({
-      message: "Вы точно хотите удалить выбранные работы?",
-      header: "Удаление",
-      icon: "pi pi-info-circle",
-      rejectLabel: "Отмена",
-      acceptLabel: "Удалить",
-      rejectClass: "p-button-secondary p-button-outlined",
-      acceptClass: "p-button-danger",
-      accept: () => {
+    confirmDelete({
+      funcAccept: () => {
         deleteSelectedIndividualWorks();
-        toast.add({
-          severity: "success",
-          summary: "Успех",
-          detail: "Записи были успешно удалены",
-          life: 2000,
-        });
       },
-      reject: () => {
+      funcReject: () => {
         selectedIndividualWorks.value = [];
       },
     });
