+++ src/views/ParentMeetings/stores/parentMeetings.js
@@ -2,13 +2,14 @@ import { API_URL, GROUP_ID } from "@/constants";
 import DialogForm from "@/utils/dialog";
 import ky from "ky";
 import { defineStore } from "pinia";
-import { useConfirm } from "primevue/useconfirm";
 import { ref } from "vue";
 import { useVuelidate } from "@vuelidate/core";
 import parentMeetingRules from "@/validators/parentMeetingRules.js";
-import { useToast } from "primevue/usetoast";
+import { useConfirmStore } from "@/stores/confirms";
 
 export const useParentMeetingsStore = defineStore("parentMeetings", () => {
+  const { confirmAdd, confirmEdit, confirmDelete } = useConfirmStore();
+
   const dialog = ref(
     new DialogForm({
       add: "Добавление протокола",
@@ -16,8 +17,6 @@ export const useParentMeetingsStore = defineStore("parentMeetings", () => {
       view: "Просмотр протокола",
     })
   );
-
-  const toast = useToast();
 
   const parentMeeting = ref({
     meettingDate: undefined,
@@ -50,29 +49,19 @@ export const useParentMeetingsStore = defineStore("parentMeetings", () => {
   const isSubmit = ref(false);
   async function addParentMetting() {
     isSubmit.value = true;
-    if (!v$.value.$invalid) {
-      const response = await ky
-        .post(`${API_URL}/parent-meetings`, {
-          json: { groupId: GROUP_ID, ...parentMeeting.value },
-        })
-        .json();
-      parentMeetings.value.push(response);
-      dialog.value.closeDialog();
-      toast.add({
-        severity: "success",
-        summary: "Успех",
-        detail: "Протокол успешно добавлен",
-        life: 2000,
-      });
-      isSubmit.value = false;
-    } else {
-      toast.add({
-        severity: "warn",
-        summary: "Предупреждение",
-        detail: "Заполните все обязательные поля",
-        life: 2000,
-      });
-    }
+    confirmAdd({
+      invalid: v$.value.$invalid,
+      funcIf: async () => {
+        const response = await ky
+          .post(`${API_URL}/parent-meetings`, {
+            json: { groupId: GROUP_ID, ...parentMeeting.value },
+          })
+          .json();
+        parentMeetings.value.push(response);
+        dialog.value.closeDialog();
+        isSubmit.value = false;
+      },
+    });
   }
 
   async function editParentMetting() {
@@ -82,38 +71,17 @@ export const useParentMeetingsStore = defineStore("parentMeetings", () => {
       })
       .json();
     await fetchParentMeetings();
-    isSubmit.value = false;
   }
 
   const confirmEditParentMeeting = () => {
-    confirn.require({
-      message: "Сохранить изменения?",
-      header: "Редактирование протокола",
-      icon: "pi pi-info-circle",
-      rejectLabel: "Отмена",
-      rejectClass: "p-button-secondary p-button-outlined",
-      acceptLabel: "Редактировать",
-      acceptClass: "p-button-info",
-      accept: () => {
-        if (!v$.value.$invalid) {
-          editParentMetting();
-          dialog.value.closeDialog();
-          toast.add({
-            severity: "info",
-            summary: "Инфо",
-            detail: "Протокол успешно изменен",
-            life: 2000,
-          });
-        } else {
-          toast.add({
-            severity: "warn",
-            summary: "Предупреждение",
-            detail: "Заполните все обязательные поля",
-            life: 2000,
-          });
-        }
+    confirmEdit({
+      invalid: v$.value.$invalid,
+      funcAccept: async () => {
+        await editParentMetting();
+        dialog.value.closeDialog();
+        isSubmit.value = false;
       },
-      reject: () => {
+      funcReject: () => {
         dialog.value.closeDialog();
       },
     });
@@ -124,31 +92,16 @@ export const useParentMeetingsStore = defineStore("parentMeetings", () => {
     parentMeetings.value = parentMeetings.value.filter(
       (meet) => meet.id !== id
     );
-    toast.add({
-      severity: "success",
-      summary: "Успех",
-      detail: "Протокол успешно удален",
-      life: 2000,
-    });
   }
 
-  const confirn = useConfirm();
-
   const confirmDeleteParentMeeting = () => {
-    confirn.require({
-      message: "Вы точно хотите удалить выбранные протоколы?",
-      header: "Удаление протокола",
-      icon: "pi pi-info-circle",
-      rejectLabel: "Отмена",
-      rejectClass: "p-button-secondary p-button-outlined",
-      acceptLabel: "Удалить",
-      acceptClass: "p-button-danger",
-      accept: () => {
+    confirmDelete({
+      funcAccept: () => {
         selectedParentMeetings.value.forEach((meet) => {
           deleteParentMeeting(meet.id);
         });
       },
-      reject: () => {
+      funcReject: () => {
         selectedParentMeetings.value = [];
       },
     });
